requirements.txt
pyyaml
openapi-spec-validator